% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mingiPolygons-package.R
\docType{package}
\name{mingiPolygons-package}
\alias{mingiPolygons}
\alias{mingiPolygons-package}
\title{mingiPolygons: Function for Identifying Polygon Objects In A Map In Shiny}
\description{
This package is able to select any polgyon object(country, county or state) from any map visualization formed by ggplot or any other package extension deployed in Shiny. It is an extension of nearPoints() but specifically for maps.
}
\author{
\strong{Maintainer}: Martin mmburu961@gmail.com Mburu

Authors:
\itemize{
  \item Ted muthomited@gmail.com Muthomi
}

}
\keyword{internal}
